initfromtextfile()successful

+++ src/Dijkstra/GraphInterface.java
@@ -11,7 +11,7 @@ public interface GraphInterface {
 	public ArrayList<VertexInterface> getSuccessors(VertexInterface vertex) ;
 	//Returns the successor of vertex
 	
-	public int getWeight(VertexInterface src,VertexInterface dst) ;
+	public double getWeight(VertexInterface src,VertexInterface dst) ;
 	//Returns the weight of the path between the vertices src and dst
 	
 	public boolean isInG(VertexInterface x);
